+++ src/main/java/com/bitbox/admin/domain/ClassAdmin.java
@@ -1,0 +1,22 @@
+package com.bitbox.admin.domain;
+
+import com.bitbox.admin.domain.key.ClassAdminId;
+
+import javax.persistence.*;
+
+@Table(name = "class_admin")
+@Entity
+public class ClassAdmin {
+    @EmbeddedId
+    private ClassAdminId id;
+
+    @MapsId("classId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "class_id")
+    private Classes classes;
+
+    @MapsId("adminId")
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "admin_id")
+    private Admin admin;
+}
