+++ src/main/java/com/bitbox/admin/domain/Classes.java
@@ -1,0 +1,40 @@
+package com.bitbox.admin.domain;
+
+import org.hibernate.annotations.DynamicInsert;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Table(name = "classes")
+@Entity
+@DynamicInsert
+public class Classes {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long classId;
+
+    @OneToMany(mappedBy = "classes", fetch = FetchType.LAZY)
+    private List<ClassAdmin> classAdmins;
+
+    @OneToMany(mappedBy = "classes", fetch = FetchType.LAZY)
+    private List<Grade> grades;
+
+    @OneToMany(mappedBy = "classes", fetch = FetchType.LAZY)
+    private List<Exam> exams;
+
+    @Column(nullable = false)
+    private String className;
+
+    @Column(nullable = false)
+    private String classCode;
+
+    @Column(nullable = false)
+    private LocalDateTime createdAt;
+
+    @Column(nullable = false, columnDefinition = "boolean default false")
+    private boolean isDeleted;
+
+    @Column(nullable = false, columnDefinition = "boolean default false")
+    private boolean isGraduate;
+}
