+++ src/main/java/com/bitbox/admin/domain/key/ClassAdminId.java
@@ -1,0 +1,10 @@
+package com.bitbox.admin.domain.key;
+
+import javax.persistence.Embeddable;
+import java.io.Serializable;
+
+@Embeddable
+public class ClassAdminId implements Serializable {
+    private Long classId;
+    private String adminId;
+}
