+++ src/main/java/com/bitbox/admin/exception/DuplicationEmailException.java
@@ -1,0 +1,5 @@
+package com.bitbox.admin.exception;
+
+public class DuplicationEmailException extends RuntimeException {
+    public DuplicationEmailException(String message) {super(message);}
+}
