+++ src/main/java/com/bitbox/admin/exception/InvalidAdminIdException.java
@@ -1,0 +1,5 @@
+package com.bitbox.admin.exception;
+
+public class InvalidAdminIdException extends RuntimeException {
+    public InvalidAdminIdException(String message) {super(message);}
+}
