+++ src/main/java/com/bitbox/admin/exception/response/ErrorResponse.java
@@ -1,0 +1,14 @@
+package com.bitbox.admin.exception.response;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+public class ErrorResponse {
+    private final String message;
+
+    @Builder
+    public ErrorResponse(String message) {
+        this.message = message;
+    }
+}
