+++ src/main/java/com/bitbox/admin/repository/ClassInfoRepository.java
@@ -1,0 +1,7 @@
+package com.bitbox.admin.repository;
+
+import com.bitbox.admin.domain.Classes;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ClassInfoRepository extends CrudRepository<Classes, Long> {
+}
