+++ src/main/java/com/bitbox/admin/repository/ExamInfoRepository.java
@@ -1,0 +1,9 @@
+package com.bitbox.admin.repository;
+
+import com.bixbox.admin.domain.Admin;
+import com.bitbox.admin.domain.Exam;
+import org.springframework.data.repository.CrudRepository;
+
+public interface ExamInfoRepository extends CrudRepository<Exam, Long> {
+    boolean existsByClasses_ClassIdAndExamNameAndDeletedIsFalse(Long classId, String examName);
+}
