--- src/main/java/com/bixbox/admin/AdminApplication.java
@@ -1,15 +1,0 @@
-package com.bixbox.admin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class AdminApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AdminApplication.class, args);
-	}
-
-}
