--- src/main/java/com/bixbox/admin/domain/ClassAdmin.java
@@ -1,22 +1,0 @@
-package com.bixbox.admin.domain;
-
-import com.bixbox.admin.domain.key.ClassAdminId;
-
-import javax.persistence.*;
-
-@Table(name = "class_admin")
-@Entity
-public class ClassAdmin {
-    @EmbeddedId
-    private ClassAdminId id;
-
-    @MapsId("classId")
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "class_id")
-    private Classes classes;
-
-    @MapsId("adminId")
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "admin_id")
-    private Admin admin;
-}
