--- src/main/java/com/bixbox/admin/domain/key/ClassAdminId.java
@@ -1,10 +1,0 @@
-package com.bixbox.admin.domain.key;
-
-import javax.persistence.Embeddable;
-import java.io.Serializable;
-
-@Embeddable
-public class ClassAdminId implements Serializable {
-    private Long classId;
-    private String adminId;
-}
