--- src/main/java/com/bixbox/admin/exception/DuplicationEmailException.java
@@ -1,5 +1,0 @@
-package com.bixbox.admin.exception;
-
-public class DuplicationEmailException extends RuntimeException {
-    public DuplicationEmailException(String message) {super(message);}
-}
