--- src/main/java/com/bixbox/admin/exception/InvalidAdminIdException.java
@@ -1,5 +1,0 @@
-package com.bixbox.admin.exception;
-
-public class InvalidAdminIdException extends RuntimeException {
-    public InvalidAdminIdException(String message) {super(message);}
-}
