--- src/main/java/com/bixbox/admin/exception/response/ErrorResponse.java
@@ -1,14 +1,0 @@
-package com.bixbox.admin.exception.response;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-public class ErrorResponse {
-    private final String message;
-
-    @Builder
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-}
