--- src/main/java/com/bixbox/admin/repository/ClassInfoRepository.java
@@ -1,6 +1,0 @@
-package com.bixbox.admin.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ClassInfoRepository extends CrudRepository<Class, Long> {
-}
