+++ src/test/java/com/bitbox/admin/AdminApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bitbox.admin;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AdminApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
