--- src/test/java/com/bixbox/admin/AdminApplicationTests.java
@@ -1,13 +1,0 @@
-package com.bixbox.admin;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AdminApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
